==> code/0_Crawling.ipynb <==
from bs4 import BeautifulSoup
import requests
import codecs

import pandas as pd
import numpy as np
import re

# the contents containing repeating / unnecessary info - these classes are excluded
exclude_class_list = [ "top-stories-promo-story__summary"]

exclude_starts_with = ["Additional reporting by"]
regex_keyword = """<meta content="(?P<keyword1>[^><\/\"]*)"\s[a-zA-Z="\-]*\sname="keywords"\s+\/+>"""

# Header info
header_info = pd.DataFrame(columns=["Id"])

# Method to crawl
df = pd.read_excel('../data/input/OnlineNewsPopularity.xlsx')
df = df[["Id", "url"]]

df_out = pd.read_excel( "../data/output/SS_Extracted_content.xlsx")

# for first time creation
#df_out = pd.DataFrame(columns=["Id", "url", "title", "html", "txt"])
out_index = 0

for index, rows in df.iterrows():
    doc_id = df.at[ index, "Id"]
    url = df.at[ index, "url"]
    
    if doc_id % 100 == 0:
        print( str(doc_id) + ", ", end='')
        df_out.to_excel( "../data/output/SS_Extracted_content.xlsx", index=False)
        
    if len( df_out.loc[ df_out["Id"] == doc_id]) > 0:
        # already processed
        continue

    resp = requests.get(url)

    if resp.status_code != 200:
        print("Issue in status code : " + str(doc_id)+ " : "+ url + " : " + str(resp.status_code))
        continue

    html_doc = resp.text
    soup = BeautifulSoup(html_doc, 'html.parser')
    content = "" # soup.title.string + "\n"

    for p in soup.select( "p"):
        text = p.get_text()

        if len( text.split()) > 1:
            if text not in content:
                is_in_exclude_list = False

                for exclude_class in exclude_class_list:
                    if p.has_attr("class") and \
                        exclude_class in p.get_attribute_list( "class"):

                        is_in_exclude_list = True
                        break

                for starts_string in exclude_starts_with: 
                    if text.startswith(starts_string):
                        is_in_exclude_list = True
                        break

                if not is_in_exclude_list:
                    content = content + text + "\n"

    title_string = soup.title.string

    df_out = df_out.append( [{"Id":doc_id, "url":url, "title":title_string, 
                              "html":html_doc, "content":content}])
    out_index += 1

print()
    
df_out.to_excel( "../data/output/SS_Extracted_content.xlsx", index=False)
