==> code/0_Recreate_Basic_Features.ipynb <==
#### Recreate Basic Features

The following features are already given in OnlineNewsPopularity data file. 
This notebook tries to recreate them as much as possible

* n_tokens_title
* n_tokens_content
* n_unique_tokens
* n_non_stop_words
* n_non_stop_unique_tokens
* num_hrefs
* num_self_hrefs
* num_imgs
* num_videos
* average_token_length
* num_keywords
* data_channel_is_lifestyle
* data_channel_is_entertainment
* data_channel_is_bus
* data_channel_is_socmed
* data_channel_is_tech
* data_channel_is_world

import pandas as pd
import numpy as np
import re
import requests

# For NLP actions
from bs4 import BeautifulSoup
from bs4 import element

import spacy
nlp = spacy.load('en_core_web_sm')

from collections import Counter

df_in = pd.read_excel('../data/output/SS_Extracted_content.xlsx')[["Id"]]
df_in = df_in[ df_in.Id < 20]
df_in.shape

doc_id = df_in.Id[13]

with open( "../data/output/html/" + str(doc_id) + ".html", "r", encoding="utf-8", errors='ignore') as html_file:
    html_doc = html_file.read()

print( len(html_doc))
print( html_doc[:100])
print( ".....")
print( html_doc[-100:])


df_out = pd.DataFrame(columns=["Id", "n_tokens_title", "n_tokens_content", "n_unique_tokens", \
                               "n_non_stop_words", "n_non_stop_unique_tokens", "num_hrefs", \
                               "num_self_hrefs", "num_imgs", "num_videos", "average_token_length", \
                               "num_keywords", "data_channel_is_lifestyle", \
                               "data_channel_is_entertainment", "data_channel_is_bus", \
                               "data_channel_is_socmed", "data_channel_is_tech", "data_channel_is_world"])
df_out = df_out.astype( {"Id":int, "n_tokens_title":int, "n_tokens_content":int, "n_unique_tokens":int, \
                         "n_non_stop_words":int, "n_non_stop_unique_tokens":int, "num_hrefs":int, \
                         "num_self_hrefs":int, "num_imgs":int, "num_videos":int, "average_token_length":int, \
                         "num_keywords":int, "data_channel_is_lifestyle":int, "data_channel_is_entertainment":int, \
                         "data_channel_is_bus":int, "data_channel_is_socmed":int, "data_channel_is_tech":int, \
                         "data_channel_is_world":int})

df_out.dtypes

soup = BeautifulSoup(html_doc, 'html.parser')

# Calculate title features
title = str( soup.title.string)
nlp_title = nlp(title)

n_tokens_title = 0
for token in nlp_title:
    n_tokens_title += 1
    
print( title)
print( "n_tokens_title : " , n_tokens_title)

### Extract content and other features

# the contents containing repeating / unnecessary info - these classes are excluded
exclude_class_list = [ "top-stories-promo-story__summary"]

exclude_starts_with = ["Additional reporting by"]
regex_keyword = """<meta content="(?P<keyword1>[^><\/\"]*)"\s[a-zA-Z="\-]*\sname="keywords".?.?>"""

content = "" # soup.title.string + "\n"

num_hrefs = 0
num_self_hrefs = 0
num_imgs = 0
num_videos = 0

for p in soup.select( "p"):
    for child in p.children:
        if isinstance( child, element.Tag):
            type_of_tag = child.name
            
            if not child.name is None:
                if child.name == "a":
                    num_hrefs += 1
                    
                    href_url = child.get('href')
                    if (not href_url is None) and ("mashable.com" in href_url.lower()):
                        num_self_hrefs += 1
                        
                if child.name == "img" or child.name == "picture":
                    num_imgs += 1
                    
                if child.name == "iframe":
                    if "youtube.com" in str( child.get("src")):
                        num_videos += 1
                    
    text = p.get_text()

    if len( text.split()) > 1:
        if text not in content:
            is_in_exclude_list = False

            for exclude_class in exclude_class_list:
                if p.has_attr("class") and \
                    exclude_class in p.get_attribute_list( "class"):

                    is_in_exclude_list = True
                    break

            for starts_string in exclude_starts_with: 
                if text.startswith(starts_string):
                    is_in_exclude_list = True
                    break

            if not is_in_exclude_list:
                content = content + text + "\n"

print( "num_hrefs : ", num_hrefs)
print( "num_self_hrefs : ", num_self_hrefs)
print( "num_imgs : ", num_imgs)
print( "num_videos : ", num_videos)

nlp_content = nlp(content)

tokens = [token.text for token in nlp_content]
                #if not token.is_stop and not token.is_punct]

n_tokens_content = len(tokens)

print( "n_tokens_content : ", n_tokens_content)

average_token_length = 0
for token in tokens:
    average_token_length += len(token)

average_token_length = average_token_length / n_tokens_content

print( "average_token_length : ", average_token_length)

count_unique_tokens = len( Counter(tokens))
n_unique_tokens = len( Counter(tokens)) / n_tokens_content
print( "n_unique_tokens :", n_unique_tokens)

tokens = [token.text for token in nlp_content if not token.is_stop and not token.is_punct]

n_non_stop_words = len(tokens) / n_tokens_content

print( "n_non_stop_words : ", n_non_stop_words)

n_non_stop_unique_tokens = len( Counter(tokens)) / count_unique_tokens
print( "n_non_stop_unique_tokens :", n_non_stop_unique_tokens)

# Keywords
regex_keyword = """<meta content="(?P<keyword1>[^><\/\"]*)"\s[a-zA-Z="\-]*\sname="keywords".?.?>"""

keywords = ""

for meta in soup.find_all( 'meta'):
    meta_text = str(meta)
    match = re.search( regex_keyword, meta_text)
    if match:
        if not match.group("keyword1") == None:
            keywords = match.group("keyword1")
            break
            
# if keywords is not found, take non-stop words from title
if keywords == "":
    tokens = [token.text for token in nlp_title if not token.is_stop and not token.is_punct]
    keywords = ", ".join( tokens)

print( keywords)

num_keywords = len( keywords.split(", "))
print( "num_keywords : ", num_keywords)

# Data channel
data_channel = ""
for p in soup.select( "hgroup"):
    if not p.get("data-channel") == None:
        data_channel = p.get("data-channel")
        
print(data_channel)

data_channel = data_channel.lower()

data_channel_is_lifestyle = 1 if data_channel == "culture" or data_channel == "" else 0
data_channel_is_entertainment = 1 if data_channel == "entertainment" else 0
data_channel_is_bus = 1 if data_channel == "business" else 0
data_channel_is_socmed = 1 if data_channel == "social-good" else 0
data_channel_is_tech = 1 if data_channel == "tech" else 0
data_channel_is_world = 1 if data_channel == "world" or data_channel == "u.s." else 0

print( "n_tokens_title",  n_tokens_title)
print( "n_tokens_content", n_tokens_content)
print( "n_unique_tokens", n_unique_tokens)
print( "n_non_stop_words", n_non_stop_words)
print( "n_non_stop_unique_tokens", n_non_stop_unique_tokens)
print( "num_hrefs", num_hrefs)
print( "num_self_hrefs", num_self_hrefs)
print( "num_imgs", num_imgs)
print( "num_videos", num_videos)
print( "average_token_length", average_token_length)
print( "num_keywords", num_keywords)
print( "data_channel_is_lifestyle", data_channel_is_lifestyle)
print( "data_channel_is_entertainment", data_channel_is_entertainment)
print( "data_channel_is_bus", data_channel_is_bus)
print( "data_channel_is_socmed", data_channel_is_socmed)
print( "data_channel_is_tech", data_channel_is_tech)
print( "data_channel_is_world", data_channel_is_world)

# Get the data_channel for all the files

df_in = pd.read_excel('../data/output/SS_Extracted_content.xlsx')[["Id"]]

df_out = pd.DataFrame(columns=["Id", "data_channel"])

for index, row in df_in.iterrows():
    doc_id = row.Id

    with open( "../data/output/html/" + str(doc_id) + ".html", "r", encoding="utf-8", errors='ignore') as html_file:
        html_doc = html_file.read()
    
    soup = BeautifulSoup(html_doc, 'html.parser')
    
    data_channel = ""
    for p in soup.select( "hgroup"):
        if not p.get("data-channel") == None:
            data_channel = p.get("data-channel")
    
    df_out.loc[index] = [doc_id, data_channel]
    
    if index % 100 == 0:
        print( index, ", ", end="")
    
df_out.to_excel('../data/output/0_data_channels.xlsx', index=False)
print()

### Validate the basic features creator class

from BasicFeatures import BasicFeaturesCreator as bfc

html_content = ""
with open( "../data/output/html/" + str(5935) + ".html", "r", encoding="utf-8", errors='ignore') as html_file:
    html_content = html_file.read()

len(html_content)

result = bfc.get_basic_features( html_content)
print( "n_tokens_title",  result.n_tokens_title)
print( "n_tokens_content", result.n_tokens_content)
print( "n_unique_tokens", result.n_unique_tokens)
print( "n_non_stop_words", result.n_non_stop_words)
print( "n_non_stop_unique_tokens", result.n_non_stop_unique_tokens)
print( "num_hrefs", result.num_hrefs)
print( "num_self_hrefs", result.num_self_hrefs)
print( "num_imgs", result.num_imgs)
print( "num_videos", result.num_videos)
print( "average_token_length", result.average_token_length)
print( "num_keywords", result.num_keywords)
print( "data_channel_is_lifestyle", result.data_channel_is_lifestyle)
print( "data_channel_is_entertainment", result.data_channel_is_entertainment)
print( "data_channel_is_bus", result.data_channel_is_bus)
print( "data_channel_is_socmed", result.data_channel_is_socmed)
print( "data_channel_is_tech", result.data_channel_is_tech)
print( "data_channel_is_world", result.data_channel_is_world)

#### Recreate Basic features of all the entries

import pandas as pd
from os import path

from BasicFeatures import BasicFeaturesCreator as bfc

df_in = pd.read_excel('../data/input/OnlineNewsPopularity.xlsx')[["Id", " shares"]]

df_out = None
out_filename = '../data/output/0_recreated_basic_features.xlsx'

if path.exists( out_filename):
    df_out = pd.read_excel( out_filename)
else:
    df_out = pd.DataFrame(columns=["Id", "n_tokens_title", "n_tokens_content", "n_unique_tokens", \
                                   "n_non_stop_words", "n_non_stop_unique_tokens", "num_hrefs", \
                                   "num_self_hrefs", "num_imgs", "num_videos", "average_token_length", \
                                   "num_keywords", "data_channel_is_lifestyle", \
                                   "data_channel_is_entertainment", "data_channel_is_bus", \
                                   "data_channel_is_socmed", "data_channel_is_tech", "data_channel_is_world", \
                                   "shares"])

df_out = df_out.astype( {"Id":int, "n_tokens_title":int, "n_tokens_content":int, "n_unique_tokens":float, \
                         "n_non_stop_words":float, "n_non_stop_unique_tokens":float, "num_hrefs":int, \
                         "num_self_hrefs":int, "num_imgs":int, "num_videos":int, "average_token_length":float, \
                         "num_keywords":int, "data_channel_is_lifestyle":int, "data_channel_is_entertainment":int, \
                         "data_channel_is_bus":int, "data_channel_is_socmed":int, "data_channel_is_tech":int, \
                         "data_channel_is_world":int, "shares":int})

for index, row in df_in.iterrows():
    doc_id = row.Id
    shares = row[" shares"]
    
    if index % 100 == 0:
        print( index, ", ", end="")
        
    if doc_id in df_out.Id:
        #already processed
        continue

    html_content = ""
    with open( "../data/output/html/" + str(doc_id) + ".html", "r", encoding="utf-8", errors='ignore') as html_file:
        html_content = html_file.read()
    
    result = bfc.get_basic_features( html_content)
    
    df_out.loc[index] = [doc_id, result.n_tokens_title, result.n_tokens_content, result.n_unique_tokens, \
                            result.n_non_stop_words, result.n_non_stop_unique_tokens, result.num_hrefs, \
                            result.num_self_hrefs, result.num_imgs, result.num_videos, result.average_token_length, \
                            result.num_keywords, result.data_channel_is_lifestyle, \
                            result.data_channel_is_entertainment, result.data_channel_is_bus, \
                            result.data_channel_is_socmed, result.data_channel_is_tech, result.data_channel_is_world, \
                            shares
                        ]
    
    if index % 100 == 0:
        df_out.to_excel('../data/output/0_recreated_basic_features.xlsx', index=False)
    
df_out.to_excel('../data/output/0_recreated_basic_features.xlsx', index=False)
print()
