==> code/2_Day_of_Week_Extraction.ipynb <==
# Libraries to remove warnings
import warnings 
warnings.filterwarnings('ignore')

# Libraries for preprocessing, scaling and model evaluation
import re
import datetime as dt
import pandas as pd


# read dataset in dataframe
df_orig = pd.read_excel('../data/input/OnlineNewsPopularity.xlsx') 
df_orig.head(5)

df_orig['date']=df_orig['url'].apply(lambda x: dt.datetime.strptime(re.search("\d{4}/\d{2}/\d{2}", x).group(), '%Y/%m/%d').date())

df_orig['date'] = pd.to_datetime(df_orig['date'])
df_orig['day_of_week'] = df_orig['date'].dt.day_name()
df_orig.head(5)

df_orig = pd.concat([df_orig, pd.get_dummies(df_orig['day_of_week'], prefix='is_')], axis=1)
df_orig.drop(['date', 'day_of_week'], axis = 1, inplace = True)
df_orig.head()

def weekend(row):
    if (row['is__Saturday'] == 1) | (row['is__Sunday'] == 1):
        return 1
    else:
        return 0

df_orig['is_Weekend'] = df_orig.apply(weekend, axis=1) 

# Display only weekend == 1 rows
df_orig[df_orig['is_Weekend'] == 1].head(100)

df_days = df_orig[ ['Id', 'is__Monday', 'is__Tuesday', 'is__Wednesday', 'is__Thursday', 'is__Friday', 'is__Saturday', 'is__Sunday', 'is_Weekend']]

df_days.head(400)

df_days.to_excel('../data/output/2_Day_of_Week_Extraction.xlsx', index=False)
