==> code/2_NLP_Clustering.ipynb <==
#Author: kannan / unsupervised - clustering of content & keywords
import pandas as pd
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.cluster import KMeans
from sklearn.metrics import adjusted_rand_score
from sklearn import metrics
import joblib

find_Opt_clusters=False

#Reading the extracted cotent
in_filepath='../data/output/SS_Extracted_content.xlsx'
out_filepath='../data/output/2_NLP_Clustering.xlsx'

df=pd.read_excel(in_filepath)
df.drop(['url','html'],axis=1,inplace=True)
df.head()

#Vectorizing the text data
documents = df['content'].tolist()

vectorizer = TfidfVectorizer(stop_words='english')
X = vectorizer.fit_transform(documents)

if find_Opt_clusters:
    #optimum number of clusters
    #Identifying optimum number of clusters\
    #https://scikit-learn.org/stable/auto_examples/cluster/plot_kmeans_silhouette_analysis.html
    c=[]
    sil=[]
    for NUM_CLUSTERS in range(5,15):
        print(NUM_CLUSTERS)
        #kmeans = cluster.KMeans(n_clusters=NUM_CLUSTERS)
        kmeans = KMeans(n_clusters=true_k, init='k-means++', max_iter=100, n_init=1)
        kmeans.fit(X)
        labels = kmeans.labels_
        silhouette_score = metrics.silhouette_score(X, labels, metric='euclidean')
        c.append(NUM_CLUSTERS)
        sil.append(silhouette_score)

    from pylab import *
    plot(c, sil)

    xlabel('#Clusters')
    ylabel('Silhouette Score')
    title('Number of Clusters')
    grid(True)
    show()

#Clustering
true_k = 10
model = KMeans(n_clusters=true_k, init='k-means++', max_iter=100, n_init=1)
model.fit(X)

#writing the clustering model
joblib.dump(vectorizer, '../data/output/models/vectorizer_Tfidf.pkl')
joblib.dump(model, '../data/output/models/Clustering.pkl')

#Visualizing clusters
print("Top terms per cluster:")
order_centroids = model.cluster_centers_.argsort()[:, ::-1]
terms = vectorizer.get_feature_names()
for i in range(true_k):
    print("Cluster %d:" % i),
    for ind in order_centroids[i, :10]:
        print(' %s' % terms[ind]),
    print

lables=model.labels_
df['cluster'] = lables.tolist()

df['cluster']='cluster_'+df['cluster'].apply(str)

df.head()

#Get one hot encoding of cluster
one_hot = pd.get_dummies(df['cluster'])
# Drop column B as it is now encoded
df = df.drop('cluster',axis = 1)
# Join the encoded df
df = df.join(one_hot)
df.head()

#New test example
test_str=df.loc[0,'content']
test_str[0:500]

df.drop(['title','content'],axis=1,inplace=True)
df.head()

#writing the output
#df.to_excel(out_filepath, index=False)

#Prediction
Y = vectorizer.transform([test_str])
prediction = model.predict(Y)
print(prediction)
