==> code/2_NLP_Keyword extraction.ipynb <==
from bs4 import BeautifulSoup
import requests
import codecs

import pandas as pd
import numpy as np
import re

regex_keyword = """<meta content="(?P<keyword1>[^><\/"]*)"\s[a-zA-Z="\-]*\sname="keywords">"""
print( regex_keyword)

# Read the html files and extract keywords
df = pd.read_excel('../data/output/SS_Extracted_content.xlsx')

df_out = pd.DataFrame(columns=['Id', 'keywords'])

for index, row in df.iterrows():
    id = row.Id
    
    html_doc = row.html
    
    soup = BeautifulSoup(html_doc, 'html.parser')
    keywords = ""
    for meta in soup.find_all( 'meta'):
        meta_text = str(meta)
        #print( meta_text)
        match = re.search( regex_keyword, meta_text)
        #print( match)
        if match:
            if not match.group("keyword1") == None:
                keywords = match.group("keyword1")
                break
                
    #print(keywords)
    df_out.loc[index] = [id, keywords]
    
    if index % 100 == 0:
        print( str(index) + ", ", end='')
        
print()

df_out.to_excel('../data/output/2_keywords_list.xlsx', index=False)

df_out.shape
