==> code/2_NLP_LDA_NMF_LSI-v2.ipynb <==
Kannan: performing decomposition and representing textual features (Title & Content) in n-components using LDA, NMF and LSI.

#Author: kannan / extracting LDA, NMF & LSI features
import pandas as pd
from sklearn.decomposition import NMF, LatentDirichletAllocation, TruncatedSVD
from sklearn.feature_extraction.text import CountVectorizer

from datetime import datetime
import joblib

#### Reading from extracted content file

in_filepath='../data/output/SS_Extracted_content.xlsx'
in_filepath_kw='../data/output/2_keywords_list.xlsx'
out_filepath='../data/output/2_NLP_LDA_NMF_LSI.xlsx'

df=pd.read_excel(in_filepath)
df_kw=pd.read_excel(in_filepath_kw)

df.head()

df_kw.head()

df=pd.merge(df,df_kw,on='Id',how='left')

df.head()

df.shape

#### feature extraction - decomposition

NUM_TOPICS=5

df.iloc[0]['title']
#df.iloc[0]['content']

titles=df['title'].values.astype('U')
titles=titles.tolist()
titles=list(filter(None,titles))
titles[0:5]

content=df['content'].values.astype('U')
content=content.tolist()
content=list(filter(None,content))
#content[0:2]

keywords=df['keywords'].values.astype('U')
keywords=keywords.tolist()
keywords=list(filter(None,keywords))
keywords[0:5]

%%time
#count vectorizing the data
#reference: https://scikit-learn.org/stable/modules/generated/sklearn.feature_extraction.text.CountVectorizer.html
vectorizerT = CountVectorizer(analyzer='word',min_df=5, max_df=0.95, 
                             stop_words='english', lowercase=True, encoding='utf-8',
                             token_pattern='[a-zA-Z\-][a-zA-Z\-]{2,}')
vectorizerC = CountVectorizer(analyzer='word',min_df=5, max_df=0.95, 
                             stop_words='english', lowercase=True, encoding='utf-8',
                             token_pattern='[a-zA-Z\-][a-zA-Z\-]{2,}')
vectorizerK = CountVectorizer(analyzer='word',min_df=5, max_df=0.95, 
                             stop_words='english', lowercase=True, encoding='utf-8',
                             token_pattern='[a-zA-Z\-][a-zA-Z\-]{2,}')

title_vectorized = vectorizerT.fit_transform(titles)
content_vectorized = vectorizerC.fit_transform(content)
keyword_vectorized = vectorizerK.fit_transform(keywords)

joblib.dump(vectorizerT, '../data/output/models/vectorizerT.pkl')
joblib.dump(vectorizerC, '../data/output/models/vectorizerC.pkl')
joblib.dump(vectorizerK, '../data/output/models/vectorizerK.pkl')

%%time
# Build a Latent Dirichlet Allocation Model
lda_modelT = LatentDirichletAllocation(n_components=NUM_TOPICS, max_iter=10, learning_method='online')
lda_modelC = LatentDirichletAllocation(n_components=NUM_TOPICS, max_iter=10, learning_method='online')
lda_modelK = LatentDirichletAllocation(n_components=NUM_TOPICS, max_iter=10, learning_method='online')
lda_title = lda_modelT.fit_transform(title_vectorized)
lda_content = lda_modelC.fit_transform(content_vectorized)
lda_keyword = lda_modelK.fit_transform(keyword_vectorized)
print(lda_title.shape)  # (NO_DOCUMENTS, NO_TOPICS)

joblib.dump(lda_modelT, '../data/output/models/lda_modelT.pkl')
joblib.dump(lda_modelC, '../data/output/models/lda_modelC.pkl')
joblib.dump(lda_modelK, '../data/output/models/lda_modelK.pkl')

%%time
# Build a Non-Negative Matrix Factorization Model
nmf_modelT = NMF(n_components=NUM_TOPICS)
nmf_modelC = NMF(n_components=NUM_TOPICS)
nmf_modelK = NMF(n_components=NUM_TOPICS)
nmf_title = nmf_modelT.fit_transform(title_vectorized)
nmf_content = nmf_modelC.fit_transform(content_vectorized)
nmf_keyword = nmf_modelK.fit_transform(keyword_vectorized)
print(nmf_title.shape)  # (NO_DOCUMENTS, NO_TOPICS)

joblib.dump(nmf_modelT, '../data/output/models/nmf_modelT.pkl')
joblib.dump(nmf_modelC, '../data/output/models/nmf_modelC.pkl')
joblib.dump(nmf_modelK, '../data/output/models/nmf_modelK.pkl')

%%time
# Build a Latent Semantic Indexing Model
lsi_modelT = TruncatedSVD(n_components=NUM_TOPICS)
lsi_modelC = TruncatedSVD(n_components=NUM_TOPICS)
lsi_modelK = TruncatedSVD(n_components=NUM_TOPICS)
lsi_title = lsi_modelT.fit_transform(title_vectorized)
lsi_content = lsi_modelC.fit_transform(content_vectorized)
lsi_keyword = lsi_modelK.fit_transform(keyword_vectorized)
print(lsi_title.shape)  # (NO_DOCUMENTS, NO_TOPICS)

joblib.dump(lsi_modelT, '../data/output/models/lsi_modelT.pkl')
joblib.dump(lsi_modelC, '../data/output/models/lsi_modelC.pkl')
joblib.dump(lsi_modelK, '../data/output/models/lsi_modelK.pkl')

models=[]
models.append(('LDA_Title',lda_title))
models.append(('NMF_Title',nmf_title))
models.append(('LSI_Title',lsi_title))
models.append(('LDA_Content',lda_content))
models.append(('NMF_Content',nmf_content))
models.append(('LSI_Content',lsi_content))
models.append(('LDA_Keyword',lda_keyword))
models.append(('NMF_Keyword',nmf_keyword))
models.append(('LSI_Keyword',lsi_keyword))

for name, model in models:
    cols=[]
    for i in range(NUM_TOPICS):
        cols.append(name[:5]+str(i))
    tmp=pd.DataFrame(model,columns=cols)
    tmp['Id']=tmp.index
    tmp['Id']=tmp.Id+1
    df=pd.merge(df,tmp,on='Id',how='left')

df.head(2)

df.columns

#writing the output
#kannan: we can remove url, title, content. just keep id and calculated columns. later we can join using 'id'
df.drop(['url', 'title','content','html','keywords'], axis=1,inplace=True)
df.to_excel(out_filepath, index=False)

"""
#Just visualizing the n_components representation of rows/observations
print(lda_title[0])
print(nmf_title[0])
print(lsi_title[0])
"""

"""
#Informational (how to extract topics)
def print_topics(model, vectorizer, top_n=10):
    for idx, topic in enumerate(model.components_):
        print("Topic %d:" % (idx))
        print([(vectorizer.get_feature_names()[i], topic[i])
                        for i in topic.argsort()[:-top_n - 1:-1]])
print("LDA Model:")
print_topics(lda_model, vectorizer)
print("=" * 20)
 
print("NMF Model:")
print_topics(nmf_model, vectorizer)
print("=" * 20)
 
print("LSI Model:")
print_topics(lsi_model, vectorizer)
print("=" * 20)
"""

"""
def infer_components(model,vectorizer,text):
    c=model.transform(vectorizer.transform([text]))[0]
    return c.tolist()

def infer_components_all(models,vectorizer,text):
    components=[]
    for _,model in models:
        components.extend(infer_components(model,vectorizer,text))
    return components
""" 

"""
text = "Kannan is a good boy. An example of new sentence"
x = lda_model.transform(vectorizer.transform([text]))[0]
print(x)
"""

"""
infer_components(lda_model,vectorizer,text)
"""
