==> code/2_NLP_NER extraction.ipynb <==
import nltk
from nltk.corpus import stopwords
from nltk.tokenize import RegexpTokenizer, sent_tokenize, word_tokenize

import numpy as np
import pandas as pd

import spacy
from spacy import displacy
from collections import Counter
import en_core_web_sm

from matplotlib import pyplot as plt

import gensim

nlp = en_core_web_sm.load()

# Read the contents
df = pd.read_excel('../data/output/SS_Extracted_content.xlsx')

df["sent_list"] = None
df["refined_content"] = None
df["NER_list"] = None
df["NER_most_common"] = None

ignored_sent_count = 0

for index, row in df.iterrows():
    sent_list = []
    
    d = str( row.content)
    d = d.replace(".\n", ". ")
    d = d.replace(".\r", ". ")
    d = d.replace("\n", ". ")
    d = d.replace("\r", ". ")
    sent = sent_tokenize(d)
    
    sent = [ s for s in sent if s != "." ] # remove sentenances with only a dot
    
    # Ignore Non-english sentenances
    # Sentenances with more than 50% of unicode chars are ignored
    for each_sent in sent:
        
        if each_sent.startswith( "{\"player\":{\"description\":"):
            continue
        
        non_english_count = 0
        for c in each_sent:
            if ord(c) > 255:
                non_english_count += 1

        if len(each_sent) > 2 and \
            non_english_count > len(each_sent)/2:

            # ignore this sentenance
            ignored_sent_count += 1
        else:
            sent_list.append( each_sent)
            
    
    refined_content = ""
    for sent in sent_list:
        if len( refined_content) > 0:
            refined_content = refined_content + " "
        
        refined_content = refined_content + sent
    
    df.at[index, 'sent_list'] = sent_list
    df.at[index, 'refined_content'] = refined_content
    
    if index % 100 == 0:
        print( str(index) + ", ", end='')
    
print( ".\n sentences extraction complete")

NER_labels = set()
for index, row in df.iterrows():
    sample_doc = df.at[index, 'refined_content']
    sample_doc = nlp( sample_doc)
    
    NER_labels.update( [x.label_ for x in sample_doc.ents])
    
    counter = Counter( [x.label_ for x in sample_doc.ents])
    
    for label in counter:
        df.at[index, "NER_" + label] = counter[label]
        
    # extract indiviudal NER entities
    counter = Counter([ent.text for ent in sample_doc.ents])
    NER_list = [text for text in counter]
    NER_most_common = [text[0] for text in counter.most_common(10)]
    
    df.at[index, "NER_list"] = ",".join(NER_list)
    df.at[index, "NER_most_common"] = ",".join(NER_most_common)
        
    if index % 100 == 0:
        print( str(index) + ", ", end='')
        
    
# set NaNs in df to 0
for label in NER_labels:
    df["NER_" + label].fillna(0, inplace=True)
        
print( ".\n NER count extraction complete")

df.count()

df_new = df.drop( 'sent_list', axis=1)

# Write all values in one fine
df_new.to_excel('../data/output/SS_Extracted_content_NER_all.xlsx', index=False)

df_NER_text = df[ ['Id', 'url', 'NER_list', 'NER_most_common']]
df_NER_text.count()

df_NER_text.to_excel('../data/output/SS_Extracted_content_NER_text.xlsx', index=False)

df_numeric = df.drop( ['url', 'html', 'title', 'content', 'refined_content', 'sent_list','NER_list', 'NER_most_common'], axis=1)
df_numeric.count()

df_numeric.to_excel('../data/output/2_NER_Type_Count.xlsx', index=False)
