==> code/2_NLP_Subjectivity_Polarity.ipynb <==
#Author: kannan
import pandas as pd
from textblob import TextBlob
from datetime import datetime

#### reading from consolidated excel file

in_filepath='../data/output/SS_Extracted_content.xlsx'
out_filepath='../data/output/2_Sentiment_Polarity_Subjectivity.xlsx'
df=pd.read_excel(in_filepath)
df.head()

df.shape

#### calculating subjectivity and 

df.iloc[0]['content']

#kannan: TextBlob tutorial: https://textblob.readthedocs.io/en/dev/quickstart.html
tmp=TextBlob(df.iloc[0]['content'])
tmp.sentiment

#kannan: If required we trip the content before calculating sentiment (e.g content=left(content, 1000))
def left(s, amount):
    return s[:amount]

def right(s, amount):
    return s[-amount:]

def mid(s, offset, amount):
    return s[offset:offset+amount]

#Iterate through all documents
df['polarity_title']=0
df['subjectivity_title']=0
df['polarity_content']=0
df['subjectivity_content']=0

startTimeModule = datetime.now()
for idx, row in df.iterrows():
    #if idx>100: #exit condition
    #    break
    title=TextBlob(str(row['title']))
    content=TextBlob(str(row['content']))
    
    df.loc[idx,'polarity_title']=title.sentiment.polarity
    df.loc[idx,'subjectivity_title']=title.sentiment.subjectivity
    
    df.loc[idx,'polarity_content']=content.sentiment.polarity
    df.loc[idx,'subjectivity_content']=content.sentiment.subjectivity

print ('Sentiment Polarity & Subjectivity extraction time:',(datetime.now() - startTimeModule))

df.head()

#writing the output
#kannan: we can remove url, title, content. just keep id and calculated columns. later we can join using 'id'
df.drop(['url', 'title','html','content'], axis=1,inplace=True)
df.to_excel(out_filepath, index=False)
