==> code/2_ONPdoc2vec_entire_content.ipynb <==
### Validate ONPdoc2vec.ONPd2v class

import numpy as np
import pandas as pd
import pickle

import string
import nltk
from nltk.corpus import stopwords

from ONPdoc2vec import ONPd2v

cached_stopwords=stopwords.words('english')

punc=string.punctuation
table=str.maketrans('','',string.punctuation)

def prep(text):
    #word tokenization
    tokens=nltk.word_tokenize(text)
    
    #converting to lower case
    tokens=[t.lower() for t in tokens]
    
    #removing punctuations
    tokens=[t.translate(table) for t in tokens]
    
    #filter tokens by length (minimum 3 chars)
    tokens=[t for t in tokens if len(t)>2]
    
    #remove stopwords
    tokens=[t for t in tokens if t not in cached_stopwords]
    
    return ",".join( tokens)

prep( "This is for testing the NLP mini project").split(",")

size = 100
pca_count = 10

filename = '../data/output/SS_Extracted_content_NER_all.xlsx'
column = 'refined_content'

print( "Processing " + filename + " " + column)
df = pd.read_excel(filename)

df["refined_content"] = df.refined_content.apply( prep)

df.at[0, 'refined_content']


d2v_model = ONPd2v( df[column].values, size=size, pca_count=pca_count)

# Save the model
d2v_model.save_model( '../data/output/models/d2v_content.model')

all_columns_list = [ "content_" + str(i) for i in range(size)]
pca_columns_list = [ "content_pca_" + str(i) for i in range(pca_count)]

df_all = df[["Id"]]
for pca_column in pca_columns_list:
    df_all[ pca_column] = 0.0

for all_column in all_columns_list:
    df_all[ all_column] = 0.0

for index, row in df_all.iterrows():
    pca_values = d2v_model.infer_vector_pca( df.at[ index, column])
    for i in range(pca_count):
        pca_column = "content_pca_" + str(i)
        df_all.at[ index, pca_column] = pca_values[i]

    all_values = d2v_model.infer_vector( df.at[ index, column])
    for i in range(size):
        all_column = "content_" + str(i)
        df_all.at[ index, all_column] = all_values[i]
        
    if index % 100 == 0:
        print( str(index) + ", ", end='')
    
print()

df_all.to_excel('../data/output/2_d2v_content.xlsx', index=False)

print( "Completed d2v model and matrix for entire content")

### for Keywords

df = pd.read_excel('../data/output/2_keywords_list.xlsx')
d2v_model = ONPd2v( df['keywords'].values, size=size, pca_count=pca_count)
column = "keywords"

# Save the model
d2v_model.save_model( '../data/output/models/d2v_keywords.model')


all_columns_list = [ "keywords_" + str(i) for i in range(size)]
pca_columns_list = [ "keywords_pca_" + str(i) for i in range(pca_count)]

df_all = df[["Id"]]
for pca_column in pca_columns_list:
    df_all[ pca_column] = 0.0

for all_column in all_columns_list:
    df_all[ all_column] = 0.0

for index, row in df_all.iterrows():
    pca_values = d2v_model.infer_vector_pca( df.at[ index, column])
    for i in range(pca_count):
        pca_column = "keywords_pca_" + str(i)
        df_all.at[ index, pca_column] = pca_values[i]

    all_values = d2v_model.infer_vector( df.at[ index, column])
    for i in range(size):
        all_column = "keywords_" + str(i)
        df_all.at[ index, all_column] = all_values[i]
        
    if index % 100 == 0:
        print( str(index) + ", ", end='')
    
print()

df_all.to_excel('../data/output/2_d2v_keywords.xlsx', index=False)

print( "Completed d2v model and matrix for all keywords")


#### NER List d2v model and output

column = "NER_list"
df = pd.read_excel('../data/output/SS_Extracted_content_NER_text.xlsx')
d2v_model = ONPd2v( df[column].values, size=size, pca_count=pca_count)

# Save the model
d2v_model.save_model( '../data/output/models/d2v_NER_list.model')

all_columns_list = [ "NER_list_" + str(i) for i in range(size)]
pca_columns_list = [ "NER_list_pca_" + str(i) for i in range(pca_count)]

df_all = df[["Id"]]
for pca_column in pca_columns_list:
    df_all[ pca_column] = 0.0

for all_column in all_columns_list:
    df_all[ all_column] = 0.0

for index, row in df_all.iterrows():
    pca_values = d2v_model.infer_vector_pca( df.at[ index, column])
    for i in range(pca_count):
        pca_column = "NER_list_pca_" + str(i)
        df_all.at[ index, pca_column] = pca_values[i]

    all_values = d2v_model.infer_vector( df.at[ index, column])
    for i in range(size):
        all_column = "NER_list_" + str(i)
        df_all.at[ index, all_column] = all_values[i]
        
    if index % 100 == 0:
        print( str(index) + ", ", end='')
    
print()

df_all.to_excel('../data/output/2_d2v_NER_list.xlsx', index=False)

print( "Completed d2v model and matrix for all NER list")

### NER most common

column = "NER_most_common"
df = pd.read_excel('../data/output/SS_Extracted_content_NER_text.xlsx')
d2v_model = ONPd2v( df[column].values, size=size, pca_count=pca_count)

# Save the model
d2v_model.save_model( '../data/output/models/d2v_NER_most_common.model')

all_columns_list = [ "NER_most_common_" + str(i) for i in range(size)]
pca_columns_list = [ "NER_most_common_pca_" + str(i) for i in range(pca_count)]

df_all = df[["Id"]]
for pca_column in pca_columns_list:
    df_all[ pca_column] = 0.0

for all_column in all_columns_list:
    df_all[ all_column] = 0.0

for index, row in df_all.iterrows():
    pca_values = d2v_model.infer_vector_pca( df.at[ index, column])
    for i in range(pca_count):
        pca_column = "NER_most_common_pca_" + str(i)
        df_all.at[ index, pca_column] = pca_values[i]

    all_values = d2v_model.infer_vector( df.at[ index, column])
    for i in range(size):
        all_column = "NER_most_common_" + str(i)
        df_all.at[ index, all_column] = all_values[i]
        
    if index % 100 == 0:
        print( str(index) + ", ", end='')
    
print()

df_all.to_excel('../data/output/2_d2v_NER_most_common.xlsx', index=False)

print( "Completed d2v model and matrix for all NER most common")
